Introduce incident specification
This commit introduces an extremely limited set of specifications:

1. The incident specification, and
2. Indident Specifications

This API is just illustrative at this time; designed to be the minimal
that is required to start setting up packaging and implementation
toolchains so that the API can be iterated on with real word
implementations.

== Design Notes
=== Incident + Impact

The incident & impact definitions have been describes as have been
experienced by the author.

There will need to be another term that describes the "post incident
investigation" and so, or that relate to it in some way.