Add a "Drone" definition file
This commit introduces the "Drone" CI/CD definition file.

This file will be used to generate, test and verify the libraries each
time there is an API change.

== Design Notes
=== Empty Pipeline

At this time, the pipeline is empty. It is introduced only to be
illustrative, as part of this commit.

=== Starlark

This commit uses the "starlark" style of pipeline, rather than the yaml
definition.

The author prefers these more dynamic style of pipelines, and plans to
use the Drone Cloud to build the content.