Introduce protobuf compilation
This commit introduces the compilation of protobuf definitions to
language specific serialization and deserialization APIs.

It further goes and introduces the gRPC client and sever definition
implementations for Golang.

== Design Notes
=== Not Packaged

This commit does not package any of the generated libraries just yet.
Rather, it just generates them to determine if the syntax is correct.

=== Package change for Go

This commit also correctly renames the package name reference for the
types in the in go.

=== No gRPC

While the original intent of this work was to include the generation of
gRPC stubs, rather than just the protobuf stubs, that requires
substantially more work and will be done on a piecemeal way.