Add limited service definitions
This commit introduces some extremely basic service defintions,
similarly to CRUD style definitions.

== Design Notes
=== Requests as messages

Each RPC has a specific input parameter which takes a set of input
messages.

This follows the pattern defined at:

https://cloud.google.com/apis/design/standard_methods#update

It is a little unusual, however the author will retain it as they do not
know as much as the authors of that content.

== Chosen Methods

The methods were chosen for their ability to fulfil a set of anticipated
use cases, such as:

- Get
- Set
- Update
- Watch
- Search

Which tend to be the common user patterns for this sort of data.