Add a timeline to the incident
All incidents have a "time" dimension, through which various events of
note are happening.

This commit adds the timeline dimensions to the incident.

== Design Notes
=== Timeline Structure

The timeline structure is designed with the anticipated rendered view
something like:

  1970-01-01 00:00 — The thing happened. ["Discovery", "Impact"]
  1970-01-01 00:01 — System ${BLAH} notified user ${BLAH}.

Where:

- 1970-01-01 00:00    The moment of the event
- The thing happened  The description of the event
- ["Discovery"]       Tags to group events together

=== Interactivity

The anticipated user model would be that users are just appending a
serives of events to a given timeline, and expecting events to be
returned in a list. That is, a timeline must always "belong" to an
incident, but may have a lifecycle that exists independently of just
that incident.

This is tricky, as it's not clear how to model this relationship
with Protobuf.