Add an id field to the event
The current design of events is that they exist outside the scope of an
incident itself. Conceptually, an event might even be associated with
multiple incidents (unlikely though this seems)

However, currently events do not have an identifier which allows them to
be addressed.

This commit addresses that by giving them one.

BREAKING CHANGE:
- Modify the incident/v1alpha1/event message format.