Automatically push code to client repository
In the initial design of these APIs, there was a single repository from
which it was expected to generate the client libraries.

The first, test library was "Go" which relies either on:

1. Not widely used proxies / metadata stores, or
2. GitHub / Git directly

The design was that the go content would sit on the "go" branch, the
"ruby" content on the "ruby" branch and so on.

However, Go quickly ran into limitations with this design. Go expects
the canonical reference to be on the master branch, use tags and other
common application versioning patterns.

Further, on reflection there is nothing particularly "bad" with this —
it is the "least surprising" implementation.

Given this, this commit instead of using the git branch pattern uses
another, separate git repository.

== Design Notes
=== Easier builds in future

This should yields easier builds in future, as each of the client
repositories can take care of their own correctness checking and
publication of libraries to the various language specific package
stores.

Additionally, examples or other metadata can be more easily stored in
patterns familiar to those languages.

=== "apis-go"

The name for the repository was changed a couple of times between
"apis-go" and "client-go".

The name "apis-go" was chosen as it isn't exactly a client, also being
the stub for the server implementation.

BREAKING CHANGE: The canonical store for Go client library has now
                 changed