Extend the the timeline with a "free text" field
Broadly, each timeline event might have references to other systems to
it — whether that's time series data systems, log systems etc.

Capturing these links to other systems is useful, as we can then reuse
these links in other display formats. However, figuring out how to store
the link is not trivial.

Rather than answer that question, this commit takes a different approach
— it allows the user to just store that (and other rich text
information) in a format of their choosing, within the text field.

These can be fairly easily tokenized and the links extracted as
required, but provides a familiar design for consumers and implementers
to reason through this content.

== Design Notes
=== Alternative: Reference style + links

The author started with an initial style which was the "referenence"
style. The message was structured:

[source,proto]
----
message Event {
    ...
    string description = 3;

    message Link {
        int id = 1;

        int URL = 2;
    }

    repeated Link links = 4;
}
----

However, figuring out the "id" to use was not trivial, and this design
requires some tokenizing of the text that ... doesn't feel natural.

BREAKING CHANGE: This breaks the v1alpha1 BC guarantee. Given this, the
                 minor version has been bumped.